cmd: estclient: cacerts: Implement a stable index suffix for -sepout
Having a stable file contents is quite important, for example,
when using configuration tools.

Establish subordinate CA index on the ascending order of NotBefore date
of the certificate (i.e. the 'oldest' subordinate CA will have the index 1
and the youngest sub CA will have the largest index).
This is not perfect, but should be good enough.

Signed-off-by: Ievgen Popovych <ievgenp@seetrue.ai>