cmd: estclient: cacerts: Support writing every CA cert to separate file
It is often required to have every CA certificate in its own file,
so instead of splitting them manually (which is clunky) - provide this option.

A new flag, `separate`, enables this mode.
It is possible to set filename prefix with `out` flag.
Files are written in a format `<prefix>-<index>.pem`,
`ca` is the default prefix, prefix can also end with `.pem`.